auth_service/requirements.txt
Flask
PyJWT
Werkzeug
redis
